requirements.txt
# torch
# torchvision
deep-sort-realtime
openvino-dev
opencv-python
click